requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
pyzbar
pandas
numpy
av
pyyaml